+++ python/rmm/rmm.py
@@ -217,6 +217,13 @@ def _make_finalizer(handle, stream):
         """
         Invoked when the MemoryPointer is freed
         """
-        return librmm.rmm_free(handle, stream)
+        librmm.rmm_free(handle, stream)
 
     return finalizer
+
+
+def _register_atexit_finalize():
+    """
+    Registers rmmFinalize() with ``std::atexit``.
+    """
+    librmm.register_atexit_finalize()