+++ tests/memory_tests.cpp
@@ -25,38 +25,44 @@
 
 cudaStream_t stream;
 
+// some useful allocation sizes
+constexpr size_t size_word{4};
+constexpr size_t size_kb {size_t{1}<<10};
+constexpr size_t size_mb {size_t{1}<<20};
+constexpr size_t size_gb {size_t{1}<<30};
+constexpr size_t size_tb {size_t{1}<<40};
+constexpr size_t size_pb {size_t{1}<<50};
+
 template <typename T>
-struct MemoryManagerTest : public ::testing::Test {
-  static rmmAllocationMode_t allocationMode() { return T::alloc_mode; }
+struct MemoryManagerUninitializedTest :
+    public ::testing::Test {};
 
-  static void SetUpTestCase() {
-    ASSERT_FALSE(rmmIsInitialized(0));
-    ASSERT_EQ(cudaSuccess, cudaStreamCreate(&stream));
-    rmmOptions_t options{};
-    options.allocation_mode = allocationMode();
-    ASSERT_SUCCESS(rmmInitialize(&options));
-    rmmOptions_t options_set;
-    // verify initialized
-    ASSERT_TRUE(rmmIsInitialized(&options_set));
-    // verify initialized options
-    ASSERT_EQ(options_set.allocation_mode, options.allocation_mode);
-    ASSERT_EQ(options_set.initial_pool_size, options.initial_pool_size);
-    ASSERT_EQ(options_set.enable_logging, options.enable_logging);
-  }
+template <typename T>
+struct MemoryManagerTest :
+    public ::testing::Test
+{
+    static rmmAllocationMode_t allocationMode() { return T::alloc_mode; }
 
-  static void TearDownTestCase() {
-    ASSERT_SUCCESS(rmmFinalize());
-    ASSERT_FALSE(rmmIsInitialized(0));
-    ASSERT_EQ(cudaSuccess, cudaStreamDestroy(stream));
-  }
+    static void SetUpTestCase() {
+        ASSERT_FALSE(rmmIsInitialized(0));
+        ASSERT_EQ( cudaSuccess, cudaStreamCreate(&stream) );
+        rmmOptions_t options {};
+        options.allocation_mode = allocationMode();
+        ASSERT_SUCCESS(rmmInitialize(&options));
+        // verify initialized
+        rmmOptions_t options_set;
+        ASSERT_TRUE(rmmIsInitialized(&options_set));
+        // verify initialized options
+        ASSERT_EQ(options_set.allocation_mode, options.allocation_mode);
+        ASSERT_EQ(options_set.initial_pool_size, options.initial_pool_size);
+        ASSERT_EQ(options_set.enable_logging, options.enable_logging);
+    }
 
-  // some useful allocation sizes
-  static const size_t size_word{4};
-  static const size_t size_kb{size_t{1} << 10};
-  static const size_t size_mb{size_t{1} << 20};
-  static const size_t size_gb{size_t{1} << 30};
-  static const size_t size_tb{size_t{1} << 40};
-  static const size_t size_pb{size_t{1} << 50};
+    static void TearDownTestCase() {
+        ASSERT_SUCCESS( rmmFinalize() );
+        ASSERT_FALSE(rmmIsInitialized(0));
+        ASSERT_EQ( cudaSuccess, cudaStreamDestroy(stream) );
+    }
 };
 
 
@@ -70,9 +76,16 @@ using allocation_modes = ::testing::Types< ModeType<CudaDefaultAllocation>,
                                            ModeType<CudaManagedMemory>,
                                            ModeType<static_cast<rmmAllocationMode_t>(PoolAllocation | CudaManagedMemory)>
                                          >;
-
-
+TYPED_TEST_CASE(MemoryManagerUninitializedTest, allocation_modes);
 TYPED_TEST_CASE(MemoryManagerTest, allocation_modes);
+
+// non-fixture tests of non-initialized RMM
+TYPED_TEST(MemoryManagerUninitializedTest, UninitializedAllocateFree) {
+    char *a = nullptr;
+    ASSERT_FAILURE(RMM_ALLOC(&a, size_kb, stream));
+    ASSERT_FAILURE(RMM_FREE(a, stream));
+}
+
 
 // Init / Finalize tests
 
@@ -81,10 +94,26 @@ TYPED_TEST(MemoryManagerTest, Initialize) {
     // Initialized in Fixture class.
     rmmOptions_t options;
     ASSERT_TRUE(rmmIsInitialized(&options));
+
+    // second initialization is no-op
+    ASSERT_NO_THROW(rmmInitialize(&options));
+    ASSERT_TRUE(rmmIsInitialized(&options));
 }
 
 TYPED_TEST(MemoryManagerTest, Finalize) {
     // Empty because handled in Fixture class.
+}
+
+TYPED_TEST(MemoryManagerTest, FreeInvalidPointer) {
+    char *a = (char*)100;
+    // TODO: no way to detect cnmem errors from RMM level,
+    // hence ASSERT_SUCCESS here
+    if (this->allocationMode() & PoolAllocation) {
+        ASSERT_SUCCESS(RMM_FREE(a, stream));
+    }
+    else {
+        ASSERT_FAILURE(RMM_FREE(a, stream));
+    }
 }
 
 // zero size tests
@@ -99,6 +128,11 @@ TYPED_TEST(MemoryManagerTest, NullPtrAllocateZeroBytes) {
     ASSERT_SUCCESS( RMM_ALLOC(p, 0, stream) );
 }
 
+TYPED_TEST(MemoryManagerTest, FreeNullPtr) {
+    ASSERT_SUCCESS( RMM_FREE(nullptr, stream) );
+}
+
+
 // Bad argument tests
 
 TYPED_TEST(MemoryManagerTest, NullPtrInvalidArgument) {
@@ -112,25 +146,25 @@ TYPED_TEST(MemoryManagerTest, NullPtrInvalidArgument) {
 
 TYPED_TEST(MemoryManagerTest, AllocateWord) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_word, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_word, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, AllocateKB) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_kb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_kb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, AllocateMB) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, AllocateGB) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_gb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_gb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
@@ -140,21 +174,21 @@ TYPED_TEST(MemoryManagerTest, AllocateTB) {
     ASSERT_SUCCESS( rmmGetInfo(&freeBefore, &totalBefore, stream) );
 
     if ((this->allocationMode() & CudaManagedMemory) ||
-        (this->size_tb < freeBefore)) {
+        (size_tb < freeBefore)) {
         // TODO investigate and fix this
-        //ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_tb, stream) );
+        //ASSERT_SUCCESS( RMM_ALLOC(&a, size_tb, stream) );
+        //ASSERT_FAILURE( RMM_FREE(a, stream) );
     }
     else {
-        ASSERT_FAILURE( RMM_ALLOC(&a, this->size_tb, stream) );
+        ASSERT_FAILURE( RMM_ALLOC(&a, size_tb, stream) );
+        ASSERT_FAILURE( RMM_FREE(a, stream) );
     }
-
-    ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, AllocateTooMuch) {
     char *a = 0;
-    ASSERT_FAILURE( RMM_ALLOC(&a, this->size_pb, stream) );
-    ASSERT_SUCCESS( RMM_FREE(a, stream) );
+    ASSERT_FAILURE( RMM_ALLOC(&a, size_pb, stream) );
+    ASSERT_FAILURE( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, FreeZero) {
@@ -165,29 +199,29 @@ TYPED_TEST(MemoryManagerTest, FreeZero) {
 /*
 TYPED_TEST(MemoryManagerTest, ReallocateSmaller) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_mb, stream) );
-    ASSERT_SUCCESS( RMM_REALLOC(&a, this->size_mb / 2, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
+    ASSERT_SUCCESS( RMM_REALLOC(&a, size_mb / 2, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, ReallocateMuchSmaller) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_gb, stream) );
-    ASSERT_SUCCESS( RMM_REALLOC(&a, this->size_kb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_gb, stream) );
+    ASSERT_SUCCESS( RMM_REALLOC(&a, size_kb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, ReallocateLarger) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_mb, stream) );
-    ASSERT_SUCCESS( RMM_REALLOC(&a, this->size_mb * 2, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
+    ASSERT_SUCCESS( RMM_REALLOC(&a, size_mb * 2, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MemoryManagerTest, ReallocateMuchLarger) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_kb, stream) );
-    ASSERT_SUCCESS( RMM_REALLOC(&a, this->size_gb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_kb, stream) );
+    ASSERT_SUCCESS( RMM_REALLOC(&a, size_gb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 */
@@ -201,8 +235,8 @@ TYPED_TEST(MemoryManagerTest, GetInfo) {
 TYPED_TEST(MemoryManagerTest, AllocationOffset) {
     char *a = nullptr, *b = nullptr;
     ptrdiff_t offset = -1;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_mb, stream) );
-    ASSERT_SUCCESS( RMM_ALLOC(&b, this->size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&b, size_mb, stream) );
 
     ASSERT_SUCCESS( rmmGetAllocationOffset(&offset, a, stream) );
     ASSERT_GE(offset, 0);
@@ -277,25 +311,25 @@ TYPED_TEST(MultiGPUMemoryManagerTest, NullPtrInvalidArgument) {
 
 TYPED_TEST(MultiGPUMemoryManagerTest, AllocateWord) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_word, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_word, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MultiGPUMemoryManagerTest, AllocateKB) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_kb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_kb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MultiGPUMemoryManagerTest, AllocateMB) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
 TYPED_TEST(MultiGPUMemoryManagerTest, AllocateGB) {
     char *a = 0;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_gb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_gb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
@@ -305,12 +339,12 @@ TYPED_TEST(MultiGPUMemoryManagerTest, AllocateTB) {
     ASSERT_SUCCESS( rmmGetInfo(&freeBefore, &totalBefore, stream) );
 
     if ((this->allocationMode() & CudaManagedMemory) || 
-        (this->size_tb < freeBefore)) {
+        (size_tb < freeBefore)) {
         // TODO investigate and fix this
-        //ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_tb, stream) );
+        //ASSERT_SUCCESS( RMM_ALLOC(&a, size_tb, stream) );
     }
     else {
-        ASSERT_FAILURE( RMM_ALLOC(&a, this->size_tb, stream) );
+        ASSERT_FAILURE( RMM_ALLOC(&a, size_tb, stream) );
     }
     
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
@@ -318,7 +352,7 @@ TYPED_TEST(MultiGPUMemoryManagerTest, AllocateTB) {
 
 TYPED_TEST(MultiGPUMemoryManagerTest, AllocateTooMuch) {
     char *a = 0;
-    ASSERT_FAILURE( RMM_ALLOC(&a, this->size_pb, stream) );
+    ASSERT_FAILURE( RMM_ALLOC(&a, size_pb, stream) );
     ASSERT_SUCCESS( RMM_FREE(a, stream) );
 }
 
@@ -336,8 +370,8 @@ TYPED_TEST(MultiGPUMemoryManagerTest, GetInfo) {
 TYPED_TEST(MultiGPUMemoryManagerTest, AllocationOffset) {
     char *a = nullptr, *b = nullptr;
     ptrdiff_t offset = -1;
-    ASSERT_SUCCESS( RMM_ALLOC(&a, this->size_mb, stream) );
-    ASSERT_SUCCESS( RMM_ALLOC(&b, this->size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&b, size_mb, stream) );
 
     ASSERT_SUCCESS( rmmGetAllocationOffset(&offset, a, stream) );
     ASSERT_GE(offset, 0);